Selective sync: Respect big folder threshold #5421
When creating a folder sync connection through the account or folder
wizard we didn't take the big folder threshold into account when
presenting the user the selective sync folder list.

Now big folders are deselected by default. And if that happens a label
is displayed that notifies the user of that fact.